return ordered fields, with empty lists and dicts for list/dict fields
