+++ lmtrex/tools/utils.py
@@ -2,7 +2,8 @@ import sys
 import traceback
 from uuid import UUID
 
-from lmtrex.common.lmconstants import APIService, ICON_API, ServiceProvider
+from lmtrex.common.lmconstants import ICON_API, ServiceProvider
+from lmtrex.common.s2n_type import S2nEndpoint
 
 # ......................................................
 def is_valid_uuid(uuid_to_test, version=4):
@@ -45,7 +46,7 @@ def get_traceback():
 def get_icon_url(provider_code, icon_status=None):
     """Add link to badge service with provider param and optionally icon_status."""
     url = None
-    if ServiceProvider.is_valid_service(provider_code, APIService.Badge['endpoint']):
+    if ServiceProvider.is_valid_service(provider_code, S2nEndpoint.Badge):
         url = '{}?provider={}'.format(ICON_API, provider_code)
         if icon_status:
             url = '{}&icon_status={}'.format(url, icon_status)
